delete exchange rates after copying the EC data dir
Add a function to delete the dirs and files without crashing the program in case it fails (this is just to be safe... in theory all the permissions should be fine)
Add a root logger to electroncash.__init__ to see the warning printed to the terminal in case there is any error.

Test Plan:
Delete electrum abc data dir, restart it, check that the newly created .electrum-abc/cache dir is empty.