fix error when starting without cmd when no_secp
In commit 00cb0b018f6edbc6a4e6f2d22ddbd509e5c8681e, I changed the way
the program defaults to 'gui' when it is called without command. This
broke this special case when secp is not available, because the function
showing the warning message relied on one of the default options set by
the 'gui' subparser.

To fix this, I'm  simply removing the piece of code that requires --relax_warnings to be set by removing this options altogether. This reverts https://github.com/Electron-Cash/Electron-Cash/pull/1305.

Alternatively I could have kept the option and just fixed it by testing
`"relaxwarn" is in cmdline_options`, but I don't like the idea of hiding annoying warnings. My plan is to hopefully figure out a way to get rid of the annoying warning by providing a proper python package for libsecp256k1, and make it a normal dependency.

Test plan:

Without libsecp256k1 built, run `./electrum-abc` and check that it
starts with a warning instead of crashing.