Scan the current screen for QR code
Add an option in the 'Pay to' context menu to scan the current screen
for a QR code.

Test plan:
Open 2 Electrum ABC wallets, display the Receive tab (showing a QR code) on one of them, go to the Send tab on the other and right click in the "Pay to" field to select the action. Verify that the address is correctly scanned from the other window.