refactor address formats handling
Make the Address format attributes descriptive strings rathers than
indices, use a dict for caching addresses instead of a list, add a
Address.toggle_address_format method to iterate/cycle through the list of
address formats.

In the GUI, prefer calling this new Address method that does not require
a boolean parameter, to allow toggling between 3 states soon.
Show "CashAddr BCH" instead of just "CashAddr".

In the configuration file, store address format as a string, not a boolean.

Test Plan:

Toggle through legacy and cashAddr format using the icon in the status
bar and also the combobox in the settings menu.