add a "Build avalanche proof" tool menu
This action open a file load dialog to build avalanche proof from a list
of stakes loaded from a json file.

This should enable exporting the stakes' metadata  (txid... etc) with a
hardware wallet on a computer connected to the internet, and then
signing the stakes to build the proof on an offline computer on which
the user can safely restore the wallet from seed phrase.