install PyInstaller from the source distribution package
PyInstaller 4.4 was released recently, and for the first time there are binary wheels for MacOS. I  suspect this is what breaks the build.

The update from PyInstaller 4.3 (.tar.gz) -> 4.4 (-py3-none-macosx_10_13_universal2.whl) is the most notable difference when comparing a build log that resulted in a working dmg file and a build log that resulted in a broken dmg.