macOS: Fix popup_widget display if running in dark mode
On macOS the "dark mode" is from the OS and the colors are auto-adapted
if using QPallette. As such, we should just treat "dark mode" for the
PopupWidget as if it were normal.

This fixes a display issue where if running the app using dark mode on
macos, the popup widget was completely unreabale.