fix tests by making the mock wallet less dumb
There are now more interactions with the wallet API, after recently removing a level of abstraction to improve the performance of the transaction building.

Some of the parameters returned by the wallet (wallet.get_txin_type) affect the transaction size.

Add a simple test that all tx respect the required maximum size rule.