fix vout serialization in COutPoint
It needs to be unsigned. In practice, this will never make a difference
because vout can never be large enough to cause this serialization to
fail, and for "small" numbers unsigned and signed integers serialize the
same.
In the Bitcoin ABC tests some vouts are generated as random data so it
highlighted this issue.