add a payout script to the proof
For now I use a simple binary blob, as we don't need to decode it.
This affects the proof by adding more data at the end, and also affects
the stake signatures because this script is now part of the proofid.

TODO: let users specify a payout address and build the p2pkh script from
that.

See https://reviews.bitcoinabc.org/D10109