add a checkpoint after the November 15th, 2021 upgrade
Added a checkpoint with the first block post upgrade (height=713661), do the same on testnet.
The VERIFICATION_BLOCK_MERKLE_ROOT is computed using the following command in the Eelctrum aBC console, after checking that the selected Electrum server is on the correct chain:

`network.synchronous_get(("blockchain.block.header", [713661, 713661]))`

Test plan:
check for a few blocks that Electrum ABC does not ban any electrum server, and that they are all updated as new blocks arrive.