don't set self.raw in JsonDb._write
This was introduced in https://github.com/Electron-Cash/Electron-Cash/pull/607 because at the time self.raw was used to determine if the wallet is encrypted.
In commit https://github.com/PiRK/ElectrumABC/commit/fcd52e90e62791871587d3d1b1182331c56567f7, a better way to remember the encryption status of a wallet was introduced, with `self._encryption_version`. Now `self.raw` is used only in `WalletStorage.__init__` when loading a wallet from file, and ` self._encryption_version` is updated in `set_password`.