Python 3.10 fixup
This is a very partial backport of https://github.com/Electron-Cash/Electron-Cash/commit/251770494ed561c1d5c7e373e0518bfe34ff9216
to make the parameter an integer when calling `QTimer.singleShot`

I don't know if the bool coercion is necessary. I will backport the rest of this commit if necessary when I get a chance to test this on Python 3.10.