do not raise BaseException
Backport of https://github.com/spesmilo/electrum/commit/7b507905841df3f685b92b2381e9ea8838f4ac78

It is also not good to raise Exception. But at least the execptions will not longer be caught by code trying to catch `GeneratorExit`, `KeyboardInterrupt` and `SystemExit`, so it is still an improvement.