make address.to_ui_string print addresses with CashAddr prefixes
We started using `to_full_ui_string` almost everywhere in https://github.com/Bitcoin-ABC/ElectrumABC/pull/61, so it makes sense to rename `to_full_ui_string` to `to_ui_string`, and use `to_ui_string_without_prefix` for the rare instances when we don't wan't the prefix.

This fixes also a bug because `to_full_ui_string` was called on a `PublicKey` object, which does not implement this method (but it has `to_ui_string`). Closes #172.