add a UI to choose the signature type
In #130, the message magic was changed with only a mention of it in the release notes, but no warning in the interface and no backward compatibility option.
This is the first step to let the user select the signature scheme, to be able to verify messages produced with older versions of Electrum ABC and other tools.