implement UInt256.deserialize as a convenience method
The existing `UInt256.unserialize` is clumsy to use because it requires a bytes string of the correct length.  With `deserialize`, we can simply pass a stream and let the method be responsible for consuming the correct number of bytes from the stream.

Keep the constructor for LimitedProofId consistent with the other UInt256 derived classes, which allows to implement `from_hex` in the base class. Add a `LimitedProofId.build` generator method to replace what was the constructor before.