raise errors when deserializing garbage data
This commits adds a `serialization.DeserializationError` and makes sure it is raised when bad data is provided. Tests are added to trigger this error.

Raising a specific error will help deal with user input, when providing a GUI to type in Proof or Delegation data.

To raise the error in the `uint256.py` module, I had to refactor the code a bit to avoid a circular dependency. Some basic data types were moved from `serialize.py` to a new `primitives.py` module, so that `serialize` no longer imports  `uint256`, so `uint256` can import `serialize.DeserializationError`.