Unsubscribe from individual addresses on wallet close
This should keep the subs count down for unused addresses when wallets
are closed.  If the refct for a scripthash drops to 0, we send
blockchain.scripthash.unsubscribe to the server.

Note that an individual address may be "seen" or used by multiple
wallets so we only unsub when a scripthash's refct drops to 0 (meaning
no wallet cares about it anymore).