Optimization: Persist the retired change addresses to wallet storage
This allows us to have very fast startup times in the happy case where
the retired change addresses have no new history.

Note that now it is possible for the wallet to not realize if it was
sent funds to a "retired" change address.  Such is life.  User can
disable the feature to revert to old behavior or they can select
"rebuild wallet history" in that corner case.

https://github.com/Electron-Cash/Electron-Cash/pull/2408/commits/b944b4e150ef3f1fc5a374518194e269fe842dac