Prefill the payout address in the avalanche proof tool
Use the current receive address to pre fill the payout address field in the Avalanche proof building dialog. The suggested address is the one that was the next receive address when the dialog was created (if a transaction is received to this address while the proof was being built, it will not be updated).

Minor improvements:
Improve typehints and variable name.

Add another reference to main_window in utxo_list, so we don't rely on weakly typed self.parent (QWidget) to access the main windows attributes.