qt high dpi: fix some text fields
This is a backport of https://github.com/spesmilo/electrum/commit/37809bed7459eb7cffff0b1f03954a6b3dc3a616
It fixes some widget size issues on high resolution displays, when widgets are sometimes too small to be conveniently used, by setting the widget width as a multiple of the character width.

Also fix the proxy password widget width in the networks dialog. It was set to use the same width as the proxy widget above, to avoid making the grid layout wider. Instead of fitting the proxy user and password widgets in individual grid cells, put them in a QHBoxLayout and fit the layout in two cells of the grid layout, so the two widgets together take the same width as the host and port of the previous line.