use a hash_160 function with fallback to local implementation
OpenSSL no longer provides ripemd160 by default, so `hashlib.new('ripemd160')` fails on Ubuntu 22.04. This caused errors when trying to send a transaction.

Use the `hash_160` wrapper from `bitcoin.py` which has a fallback to the local implementation in `ripemd.py` in case `hashlib` fails.

Note that this has a beneficial side effect of removing a bunch of “Shadows name from outer scope” warnings in `address.py`, as `hash160` was used as a function name, and a class member name, and a parameter name in methods.

Test plan:
Send a transaction with a standard wallet and with a ledger wallet.
TODO: test that CashFusion still works afterwards (it uses `util.tx_from_components`)