move get_public_keys to correct wallet classes
This is only implement for deterministic wallets, and the base implementation only worked for Multisig wallets. Move the method as an abstract method to DeterministicWallet, move implementations to where it makes sense, add a documentation to mention the format of the output (which does not make much sense, as a bytes string would be half as costly).