fix daemon command output
A daemon command should not print, as the body of the function will be called in the `daemon start` process which might no longer be attached to a terminal (causing an `OSError` on `print`).

Instead, the command should return the output, which is then printed in the actual process calling the command.