disable setting tx locktime to current block by default
Use a locktime=0 by default.
The reason for setting locktime to current block was to do the same as the node, but this has changed in https://reviews.bitcoinabc.org/D13059

On the eCash blockchain, fee-sniping is mitigated by post-consensus, so setting the locktime is no longer necessary. It is easier for light wallets such as CashTab to set the locktime to 0, and it is better if all wallets use the same behavior for privacy (reducing wallet fingerprinting).