fix -wp / --walletpassword option
This option was lost in a backported commit from Electrum: 4831972759cfd3e9f5f6eb15454f2da607809328
The option was initially added in Electron Cash after the fork from Electrum: https://github.com/Electron-Cash/Electron-Cash/pull/740

Note that the option must be passed to commands that need it directly ( `load_wallet`). It is not remembered between commands, so passing it on `daemon start` does not work.