bump protobuf and regenerate pb2.py files
See https://github.com/spesmilo/electrum/issues/7922 for a good description of the problem and the rational for restricting to this range of `protobuf` versions.
tldr; keepkey support restricts us to `protobuf<4` because we need to support both the new and the old format.

This fixes  https://github.com/Bitcoin-ABC/ElectrumABC/issues/214

This removes a lot of DeprecationWarning messages due to outdated pb2 files.