[electrum] convert README to markdown
Summary:
The RST format is not automatically converted into a HTML documentation page  on the bitcoinabc.org website.
Also, it makes sense to not use multiple doc formats.

Remove the redundant MIT license, authors and language at the top.

Test Plan: I tested all the URL and rendering in my IDE. It wasn't able to make the internal link (to sections in the same doc) work, but it offered autocompletion, so I'm confident that they are right.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: Fabien, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14250