apply same changes as in the monorepo integration commit
When ElectrumABC was migrated into the monorepo, a few changes had to be made to please the linter and remove unnecessary files (pre-commit and cirrus related configs).
This brings this repository into the same shape as the monorepo's `electrum` ssubdirectory right after the migration commit, so that we can then keep this repository as a mirror by applying the same commits as done in the monorepo.