[electrum] don't return dict in Transaction.serialize()
Summary:
This dict is only used in one command and 3 tests. It can be constructed where needed, and the tests can be adjusted.

I removed the test for `"prevout_n"` in the outputs. This was basically just testing that the first element in a list has index 0.

Test Plan:
```
python test_runner.py
./electrum-abc deserialize 010000000149f35e43fefd22d8bb9e4b3ff294c6286154c25712baf6ab77b646e5074d6aed010000005701ff4c53ff0488b21e0000000000000000004f130d773e678a58366711837ec2e33ea601858262f8eaef246a7ebd19909c9a03c3b30e38ca7d797fee1223df1c9827b2a9f3379768f520910260220e0560014600002300feffffffd8e43201000000000118e43201000000001976a914e158fb15c888037fdc40fb9133b4c1c3c688706488ac5fbd0700
```

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14470