[electrum] check output type in Transaction and fix a callsites that still passes a regular tuple
Summary:
As of D14448, the transaction size estimation requires outputs to be a TxOutput. As of D14502, the tx serializaation also requires a TxOuput.
Enforce this with assertions in the various output setters in `Transaction`, add typehints to functions that call these setters.

Fix a callsite in the fusion plugin that was still passing a regular tuple.

Test Plan: `python test_runner.py`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14508