[electrum] move get*Filename wrappers out of MainWindow
Summary:
The only thing these wrappers need from the MainWindow is the config, so pass it as a parameter.

Depends on D14537

Test Plan:
In the contacts tab, export contacts to a file. Try to import that file in another wallet.

Double-click on a transaction in the History tab, and click the save button.

Open a transaction from file (Tools > Load Transaction > From file)

Add a few labels to transactions in the History tab, export them (Wallet > Labels > Export), delete some of them again, then reimport and check they are all back

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14538