[electrum] use a WaitingDialog to resolve aliases
Summary:
This informs the user that something is happening, the same way we do when signing or broadcasting a transaction.

Depends on D14457

ref T3298

Test Plan:
`./electrum-abc -v --enable-aliases`

Try typing a good alias in the Pay To widget (`twelvechar12.xec`). Check that it is added to the Contacts tab, and the transaction works.
Try a random unregistered alias. Watch the terminal output for the error message.

Reviewers: #bitcoin_abc, Fabien, bytesofman

Reviewed By: #bitcoin_abc, Fabien, bytesofman

Subscribers: Fabien

Maniphest Tasks: T3298

Differential Revision: https://reviews.bitcoinabc.org/D14550