[electrum] implement removal of coins in the proof editor
Summary:
Allow removal of coins from the proof editor tool, using either a button at the end of the line or a context menu (allows deleting multiple coins at once)

This has been requested a lot, because the alternative is to manually delete the coins from the JSON file before loading it.

Depends on D14760

Test Plan: Open the proof editor with multiple coins, try the various methods for deleting coins: the red cross button, the context menu with single selection or multiple selection.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14761