[electrum] log: only log in verbose mode in CachedWalletPasswordWidget
Summary:
My intention was to log this message only in verbose mode, but it turns out `print_msg` always logs the message. We need to use `print_error` instead. This is a bit counterintuitive to me, as in other logging frameworks the severity of errors is higher than the severity of info messages and debug messages.

This antipattern of using `print_error` for debbuging logs will be cleaned out of the codebase when I eventually backport https://github.com/spesmilo/electrum/commit/3385a94753af2d93913ee9e9d67d18627c9c8f9e

Test Plan: repeat the test plan from D14729 with and without the `-v` command line option, verify that the message is only printed in verbose mode

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14765