[electrum] fix delegator key detection when building multi-level delegation
Summary: This was not updated in D14603, and as a result the delegator key would not be automatically filled when pasting an existing delegation in the delegation editor  to add another level to it.

Test Plan:
Generate a delegation, using a suggested key as the delegated key. Copy the delegation, close the delegation editor, reopen it, and paste the delegation in the "from exisitng delegation" field. Check that the delegator key is correctly set to the previous delegated key. Add another level with a new suggested key, repeat the test.

`python test_runner.py`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14802