[electrum] disable generate key button in the delegation editor for non-HD and watch-only wallets
Summary:
No need to show a live button that does nothing. To generate a key, the wallet must be a HD wallet and have access to the xpriv key.
Add a tooltip to explain why the button is grey and unresponsive.

Test Plan: Open the delegation editor with various types of wallets. Check that the button  can only be used if the wallet can generate keys.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15013