[electrum] run isort on python scripts with no extension
Summary:
Run the ISort linter on python code that was not linted because it was missing the .py extension. In a following commit, we will rename them so they will be linted automatically.
The main entry point for the application will not get a .py extension, so add it to include in .arclint.

Depends on D15357

Test Plan:
```
arc lint -- electrum/electrum-abc
```

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15358