[electrum] set dark-mode automatically for ShowPopupLabel
Summary:
There are currently two callsites for this function, and both set this flag in the same way. So just set it in the function so callers don't have to care about it.

`ColorScheme.dark_scheme` is a boolean, `True` if the user set the appeareance to Dark in the general settings.

Test Plan:
In electrumabc/version.py, decrease the version number (`VERSION_TUPLE`). Run Electrum ABC from source (`./electrum-abc`), wait for a few seconds, check that the update notification popup shows correctly in the status bar.

I tested it both in default mode and dark mode (menu `Tools > Preferences > Color Theme`, restart Electrum ABC).

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15384