[cmake] don't include check-electrum in check-all target
Summary:
All other `check-all` dependencies are more or less directly related to node development. Including `check-electrum` adds a bunch of unrelated requirements for python dependencies for node developers. This is not clearly explained in the CONTRIBUTING and build documentation, which implies that the electrum dependencies are only needed for electrum devs.

Let's remove that dependency, and run the check-electrum target only on diffs that touch the electrum subtree and on master commits. There is currently no way a commit that does not touch the electrum subtree can break the electrum unit tests, they are only related to the electrum codebase.

Test Plan:
Check that electrum unit tests are not run on `check-all`.
```
mkdir build
cd build
cmake .. -GNinja
ninja all check-all
ninja check-electrum
```

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15544