[electrum] prefill the delegator key in the delegation editor when possible
Summary:
We already prefill that key when pasting a delegation and its delegated key is recognized as an auxiliary key from this wallet.
Do it also when pasting a proof and its master pubkey is recognized.

Test Plan:
Open the delegation editor in a wallet from which a key was used to produce a proof. Paste the proof in the "From proof" tab. Check that the delegator priv key field is prefilled.

Check also that the feature still works when pasting a delegation.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15708