[electrum] bump pyinstaller to v6.10.0
Summary:
We recently bumped setuptools to >= 70.0.0, which forces us to bump pyinstaller (see https://github.com/pyinstaller/pyinstaller/pull/8555)
Fixes an error when attempting to run the generatad windows release.

Unfortunately it does not seem enough to fix the same issue on MacOS, so we also revert the version bump for setuptools for the MacOS build (partial revert of D16481), which fixes the issue.
AFAICT we are not affected by CVE-2024-6345 because we don't use the `package_index` module of setuptools, so it should be safe to use setuptools < 70.0.

Depends on D16725

Test Plan: Build the MacOS and Windows releases, then check that the executables actually work

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16726